requirements.txt

fastapi[standard]
celery
types-redis
redis
celery-types
uvicorn-worker
uvicorn[standard]
gunicorn
pydantic
ruff
flake8
pytest
pytest-cov
bandit
sqlalchemy
alembic
psycopg2-binary
python-dotenv
python-logging-loki
pyjwt
polars
prefect
prometheus-client
grafana-api
pipdeptree
httpx
structlog
importlib
rich==13.7.1
